ADOBE1A1/ADOBE1A/requirements.txt
pandas
xgboost
scikit-learn
joblib
pdfminer.six